# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Cykloservis API
  version: 0.1.0
  description: Základní veřejné a autentizační endpointy.
servers:
  - url: http://localhost:5001
paths:
...
  /service-requests:
    get:
      summary: Výpis servisních požadavků
      operationId: serviceRequestsList
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Seznam požadavků
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceRequest'
        '401':
          description: Neautorizováno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Vytvoření servisního požadavku
      operationId: serviceRequestsCreate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bikeId, title]
              properties:
                bikeId: { type: string }
                title: { type: string }
                description: { type: string }
                preferredDate: { type: string, format: date-time }
                priceEstimate: { type: number, minimum: 0 }
      responses:
        '201':
          description: Vytvořeno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRequest'
        '400':
          description: Neplatný vstup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Neautorizováno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /service-requests/{id}/status:
    put:
      summary: Aktualizace statusu požadavku
      operationId: serviceRequestsUpdateStatus
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [new, in_progress, done, cancelled] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRequest'
        '400':
          description: Neplatný vstup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Neautorizováno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /service-requests/{id}:
    delete:
      summary: Smazání servisního požadavku
      operationId: serviceRequestsDelete
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                required: [ok]
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Neautorizováno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message: { type: string }
      required: [message]
    ServiceRequest:
      type: object
      properties:
        id: { type: string, description: 'ID z Mongo ObjectId (transformováno z _id)' }
        ownerEmail: { type: string, format: email }
        bikeId: { type: string }
        title: { type: string }
        description: { type: string }
        status: { type: string, enum: [new, in_progress, done, cancelled] }
        preferredDate: { type: string, format: date-time }
        priceEstimate: { type: number }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, ownerEmail, bikeId, title, status]
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        code: { type: string }
        details: { type: object, additionalProperties: true }
      required: [error, code]
    UnauthorizedResponse:
      type: object
      properties:
        message: { type: string, example: Unauthorized }
      required: [message]
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT